==> 2_data-wide.Rmd <==
---
title: "data import"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
---

# libraries

```{r setup, echo=FALSE}
# load libraries
library(knitr)      # use .Rmd files

  # reading data
library(data.table) # fast data manipulation

  # manipulate data
library(tidyr)      # manipulating data
library(dplyr)      # manipulating data
library(rlist)      # working with lists
library(stringr)    # work with strings
library(tibble)     # create data frames
library(stringi)    # process charactar strings
library(lubridate)  # transform character into dates
library(epitools)   # time manipulation
library(lubridate)  # time manipulation (quarters)

  # statistics
library(Matrix)     # required for lme4

  # visuals
library(ggplot2)    # make graphs
library(ggExtra)    # additional functionality for ggplot2
library(ggpubr)     # combine multiple graphs into one
library(gridExtra)  # table management
```

# data import

reading in the data enginereed file.

```{r clean_data}
data <- fread("./data/data_clean.csv")
```

# wide format

generate data set for each Location to transform it into a wide format, since clustering methods use wide formats.

```{r tranform, eval=T, echo=FALSE}
# Hoepfner
Hoepfner <- data[,c(1,2)] %>% rename(
  "Leistung" = "Hoepfner"
) %>% drop_na() %>%
  distinct(Zeit_deutsch, .keep_all = TRUE)

Hoepfner$time <- str_sub(Hoepfner$Zeit_deutsch, -8, -4)
Hoepfner$date <- str_sub(Hoepfner$Zeit_deutsch, 1, 10)
Hoepfner$Zeit_deutsch <- NULL
Hoepfner <- Hoepfner[-1:-3,] # delete first rows to start at 00:00
  # transpose the data frame
Hoepfner_wide <- Hoepfner %>% pivot_wider(
  names_from = time,
  values_from = Leistung
) %>% drop_na()


# FZI
FZI <- data[,c(1,4)] %>% rename(
  "Leistung" = "FZI"
) %>% drop_na() %>%
  distinct(Zeit_deutsch, .keep_all = TRUE)

FZI$time <- str_sub(FZI$Zeit_deutsch, -8, -4)
FZI$date <- str_sub(FZI$Zeit_deutsch, 1, 10)
FZI$Zeit_deutsch <- NULL
FZI <- FZI[-1:-8,] # delete first rows to start at 00:00
  # transpose the data frame
FZI_wide <- FZI %>% pivot_wider(
  names_from = time,
  values_from = Leistung
) %>% drop_na()


# MGH
MGH <- data[,c(1,3)] %>% rename(
  "Leistung" = "MGH"
) %>% drop_na() %>%
  distinct(Zeit_deutsch, .keep_all = TRUE)

MGH$time <- str_sub(MGH$Zeit_deutsch, -8, -4)
MGH$date <- str_sub(MGH$Zeit_deutsch, 1, 10)
MGH$Zeit_deutsch <- NULL
MGH <- MGH[-1:-41,] # delete first rows to start at 00:00
  # transpose the data frame
MGH_wide <- MGH %>% pivot_wider(
  names_from = time,
  values_from = Leistung
) %>% drop_na()


# TeFak
TeFak <- data[,c(1,5)] %>% rename(
  "Leistung" = "TeFak"
) %>% drop_na() %>%
  distinct(Zeit_deutsch, .keep_all = TRUE)

TeFak$time <- str_sub(TeFak$Zeit_deutsch, -8, -4)
TeFak$date <- str_sub(TeFak$Zeit_deutsch, 1, 10)
TeFak$Zeit_deutsch <- NULL
TeFak <- TeFak[-1:-4,] # delete first rows to start at 00:00
  # transpose the data frame
TeFak_wide <- TeFak %>% pivot_wider(
  names_from = time,
  values_from = Leistung
) %>% drop_na()


# create data-wide to build time series there
  # add columns indicating location
FZI_wide$location <- 'FZI'
Hoepfner_wide$location <- 'Hoepfner'
MGH_wide$location <- 'MGH'
TeFak_wide$location <- 'TechnologieFabrik'

data_wide <- rbind(
  FZI_wide,
  Hoepfner_wide,
  MGH_wide,
  TeFak_wide
)
data_wide <- data_wide %>%
  drop_na()

write_csv(data_wide,
          file = "./data/data_wide.csv")
```

create graphs for each set of data on a daily basis

```{r visualisation, eval=T, echo=FALSE}
data$time <- as.factor(data$time)
  # Technologie Fabrik
TechnoPlot <- ggplot(
  data[,.( Avg = mean(TeFak, na.rm = TRUE)), by = time],
  aes(y = Avg, x = time)
) +
  geom_point() +
  theme_bw() +
  theme(axis.text.x = element_blank()) +
  labs(y = "Avg Power in W", x = "") +
  ggtitle("TeFak")
  # MGH
MGHPlot <- ggplot(
  data[,.( Avg = mean(MGH, na.rm = TRUE)), by = time],
  aes(y = Avg, x = time)
) +
  geom_point() +
  theme_bw() +
  theme(axis.text.x = element_blank()) +
  labs(y = "", x = "") +
  ggtitle("MGH")
  # Hoepfner
HoepfnerPlot <- ggplot(
  data[,.( Avg = mean(Hoepfner, na.rm = TRUE)), by = time],
  aes(y = Avg, x = factor(time))
) +
  geom_point() +
  theme_bw() +
  labs(y = "Avg Power in W", x = "time of day") +
  theme(axis.text.x = element_blank()) +
  ggtitle("Hoepfner Areal")
  # FZI
FZIPlot <- ggplot(
  data[,.( Avg = mean(FZI, na.rm = TRUE)), by = time],
  aes(y = Avg, x = time)
) +
  geom_point() +
  theme_bw() +
  labs(y = "", x = "time of day") +
  theme(axis.text.x = element_blank()) +
  ggtitle("FZI")

ggsave(filename = "Durchschnitssverbraeuche.png",
       device = "png",
       plot = ggarrange(TechnoPlot, MGHPlot, HoepfnerPlot, FZIPlot,
          ncol = 2, nrow = 2),
       path = "./graphs/",
       dpi = 320,
       width = 2500,
       height = 1600,
       units = "px"
)
```

# Datenauswertung

Ziel ist die Dategrundlage zu analysieren und mit Hilfe von deskriptiven Statistiken die vorhandenen Daten zu beschreiben.

```{r datenauswertung, echo=FALSE}
# Zeitstrahl für fehlende Daten
  # subset von Daten für fehlende Werte
FZI$location <- 'FZI'
Hoepfner$location <- 'Hoepfner'
MGH$location <- 'MGH'
TeFak$location <- 'TechnologieFabrik'
data_clean <- rbind(
  FZI,
  Hoepfner,
  MGH,
  TeFak
)

# create data frame to store measured time values
NA_df <- data.frame(matrix(NA,
                               nrow = 6,  # as many rows as chunks
                               ncol = 2))  # 3 columns for chunk, original and improved
NA_df <- dplyr::rename(
    NA_df,
    'Data Set/Location' = X1,
    'Number of Missing Values' = X2
  )
# name the tasks
NA_df[,1] <- c(
  "complete data set for available dates",
  "complete data set for all dates",
  "FZI",
  "Hoepfner",
  "Mehrgeneration Haus",
  "Technologie Fabrik"
)

# complete data set with all dates for every location
NA_df[1,2] <- sum(is.na(data_clean$Leistung))
# complete data set with only available dates
NA_df[2,2] <- sum(is.na(data))

# FZI standalone
NA_df[3,2] <- sum(is.na(FZI))
# Hoepfner Brauerei standalone
NA_df[4,2] <- sum(is.na(Hoepfner))
# Mehrgenerationenhaus standalone
NA_df[5,2] <- sum(is.na(MGH))
# Technologiefabrik standalone
NA_df[6,2] <- sum(is.na(TeFak))

NA_df

# table for gantt chart
gantt_df <- data.frame(matrix(NA,
                               nrow = 4,  # as many rows as buildings
                               ncol = 3))  # 3 columns for building, start and end
gantt_df <- dplyr::rename(
    gantt_df,
    'Location' = X1,
    'start' = X2,
    'end' = X3
  )
# name the tasks
gantt_df[,1] <- c(
  "TeFak",
  "MGH",
  "FZI",
  "Hoepfner"
)


for (i in 1:nrow(gantt_df)){
  gantt_df[i,2] <- min(data[!is.na(data[[i]]), date])
  gantt_df[i,3] <- max(data[!is.na(data[[i]]), date])
}
# column class as Date
gantt_df[,-1] <- lapply(gantt_df[,-1],as.Date,origin = "1970-01-01")
# Gantt chart
gantt_plot <- ggplot(gantt_df,
       aes(
         x = start,
         xend = end,
         y = Location,
         yend = Location
       )) +
  theme_bw() +
  geom_segment(size = 10, color = 'chartreuse4') +
  labs(title = "Datenverfügbarkeit nach Gebäuden",
       x = "Datum",
       y = "Gebäude")

ggsave(filename = "Datenverfügbarkeit.png",
       device = "png",
       plot = gantt_plot,
       path = "./graphs/"
)
```
check for missing values

All missing values are occurring for the Hoepfner Brauerei.
